requirements.txt
python-telegram-bot==11.1.0
psycopg2==2.8.3
